backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
torch
transformers
sentencepiece
accelerate
tf-keras
ffmpeg-python
whisperx
